Refactored oklog into oklog-core, added oklog3 with OkHttp3 support
