+++ core/src/main/java/com/github/simonpercic/oklog/core/utils/Constants.java
@@ -1,0 +1,18 @@
+package com.github.simonpercic.oklog.core.utils;
+
+/**
+ * Constants.
+ *
+ * @author Simon Percic <a href="https://github.com/simonpercic">https://github.com/simonpercic</a>
+ */
+public final class Constants {
+
+    private Constants() {
+
+    }
+
+    public static final String LOG_TAG = "OKLOG";
+    public static final String LOG_URL_BASE_LOCAL = "http://localhost:8080";
+    public static final String LOG_URL_BASE_REMOTE = "http://responseecho-simonpercic.rhcloud.com";
+    public static final String LOG_URL_ECHO_RESPONSE_PATH = "/re/";
+}
