+++ oklog3/gradle.properties
@@ -16,12 +16,6 @@
 # This option should only be used with decoupled projects. More details, visit
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
-# VERSION_CODE = XX.YYY.ZZZ
-VERSION_NAME = 0.1.3-SNAPSHOT
-VERSION_CODE = 00001003
-SITE_URL = https://github.com/simonpercic/OkLog
-GIT_URL = https://github.com/simonpercic/OkLog.git
-GROUP_ID = com.github.simonpercic
-DEVELOPER_ID = simonpercic
-DEVELOPER_NAME = Simon Percic
-DEVELOPER_EMAIL = simon@percic.si
+LIBRARY_NAME = OkLog3
+LIBRARY_DESCRIPTION = Response logging interceptor for OkHttp3. Logs a url link with url-encoded response for every OkHttp call.
+BINTRAY_NAME = oklog3
